Fix more errors with types

+++ dedup/scan.py
@@ -28,7 +28,7 @@ CREATE TABLE IF NOT EXISTS files (
     kind INTEGER,
     path TEXT,
     size INTEGER,
-    timestamp INTEGER,
+    timestamp REAL,
     checksum TEXT
 );
 """
